+++ packages/aeria-sdk/src/cli.ts
@@ -1,4 +1,4 @@
-import { parseArgs } from 'util'
+import { parseArgs } from 'node:util'
 import { mirrorRemotely } from './mirror.js'
 import { getConfig } from './config.js'