+++ packages/builtins/src/collections/file/remove.ts
@@ -2,7 +2,7 @@ import type { Context, SchemaWithId, RemovePayload } from '@aeriajs/types'
 import type { description } from './description.js'
 import { Result } from '@aeriajs/types'
 import { remove as originalRemove, get, type ObjectId } from '@aeriajs/core'
-import * as fs from 'fs/promises'
+import * as fs from 'node:fs/promises'
 
 export const remove = async (payload: RemovePayload<SchemaWithId<typeof description>>, context: Context<typeof description>) => {
   const { error, result: file } = await get({