+++ packages/compiler/src/codegen/generateExports.ts
@@ -2,8 +2,8 @@ import type * as AST from '../ast'
 import { resizeFirstChar, getExtendName } from './utils'
 
 type SymbolToExport = {
-  id: string,
-  schema: string,
+  id: string
+  schema: string
   extend: string
 }
 
@@ -11,19 +11,17 @@ type SymbolToExport = {
 const codeGenerators = new Map<string, (symbols: SymbolToExport[]) => string>([
   [
     '/out/collections/index.js',
-    (symbols: SymbolToExport[]) => `export { ${symbols.map((symbol) => `${symbol.id}`).join(', ')} } from './collections.js'`,
+    (symbols) => `export { ${symbols.map((symbol) => `${symbol.id}`).join(', ')} } from './collections.js'`,
   ],
-
   [
     '/out/index.d.ts',
-    (symbols: SymbolToExport[]) =>
+    (symbols) =>
       'export * as collections from \'./collections/index.js\'\n' +
             `export { ${symbols.map((symbol) => `${symbol.extend}, ${symbol.schema}`).join(', ')} } from './collections/collections.js'`,
   ],
-
   [
     '/out/index.js',
-    (symbols: SymbolToExport[]) =>
+    (symbols) =>
       'export * as collections from \'./collections/index.js\'\n' +
             `export { ${symbols.map((symbol) => symbol.extend).join(', ')} } from './collections/collections.js'`,
   ],
@@ -43,3 +41,4 @@ export const generateExports = (ast: AST.Node[]) => {
 
   return statements
 }
+