+++ packages/core/src/__scripts__/postinstall.ts
@@ -1,5 +1,5 @@
-import * as path from 'path'
-import * as fs from 'fs'
+import * as path from 'node:path'
+import * as fs from 'node:fs'
 
 const DTS_FILENAME = 'aeria.d.ts'