+++ packages/cli/CHANGELOG.md
@@ -1,5 +1,20 @@
 # aeria-build
 
+## 0.0.164
+
+### Patch Changes
+
+- @aeriajs/builtins@0.0.216
+- @aeriajs/core@0.0.216
+
+## 0.0.163
+
+### Patch Changes
+
+- Updated dependencies [97e2027]
+  - @aeriajs/core@0.0.215
+  - @aeriajs/builtins@0.0.215
+
 ## 0.0.162
 
 ### Patch Changes