+++ packages/core/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @aeriajs/api
 
+## 0.0.216
+
+### Patch Changes
+
+- Updated dependencies [1110583]
+  - @aeriajs/validation@0.0.140
+  - @aeriajs/builtins@0.0.216
+  - @aeriajs/http@0.0.151
+  - @aeriajs/security@0.0.216
+
+## 0.0.215
+
+### Patch Changes
+
+- 97e2027: Add base64 upload format
+  - @aeriajs/builtins@0.0.215
+  - @aeriajs/security@0.0.215
+
 ## 0.0.214
 
 ### Patch Changes