+++ packages/security/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @aeriajs/security
 
+## 0.0.216
+
+### Patch Changes
+
+- @aeriajs/core@0.0.216
+
+## 0.0.215
+
+### Patch Changes
+
+- Updated dependencies [97e2027]
+  - @aeriajs/core@0.0.215
+
 ## 0.0.214
 
 ### Patch Changes